+++ src/SettingsLoader/AbstractFileSystemSettingsStorageHandler.php
@@ -1,6 +1,5 @@
 <?php declare(strict_types=1);
 
-
 namespace Cspray\Labrador\SettingsLoader;
 
 use Cspray\Labrador\Exception\InvalidArgumentException;