+++ src/SettingsLoader/DefaultsWithEnvironmentOverrideSettingsLoader.php
@@ -8,7 +8,7 @@ use Cspray\Labrador\Environment;
 use Cspray\Labrador\Settings;
 
 /**
- * Loads settings at a specific file path and then overrides any of those settings based on the ApplicationEnvironment
+ * Loads settings at a specific file path and then overrides any of those settings based on the EnvironmentType
  * the Application is currently running in.
  *
  * @package Cspray\Labrador\SettingsLoader
@@ -17,20 +17,33 @@ use Cspray\Labrador\Settings;
 final class DefaultsWithEnvironmentOverrideSettingsLoader implements SettingsLoader {
 
     private $storageHandler;
-    private $settingsConfiguration;
+    private $settingsFile;
+    private $envDir;
+    private $supportedFileTypes = [];
 
-    public function __construct(SettingsStorageHandler $storageHandler, EnvironmentSettingsConfiguration $environmentSettingsConfiguration) {
+    public function __construct(
+        SettingsStorageHandler $storageHandler,
+        string $settingsFile,
+        string $envDir,
+        array $supportedFileTypes = ['php', 'json']
+    ) {
         $this->storageHandler = $storageHandler;
-        $this->settingsConfiguration = $environmentSettingsConfiguration;
+        $this->settingsFile = $settingsFile;
+        $this->envDir = $envDir;
+        $this->supportedFileTypes = $supportedFileTypes;
     }
 
     public function loadSettings(Environment $environment) : Settings {
-        $settings = new Dot($this->storageHandler->loadSettings($this->settingsConfiguration->getDefaultPath()));
-        $envPath = $this->settingsConfiguration->getPathForApplicationEnvironment($environment->getApplicationEnvironment());
-        if (isset($envPath)) {
-            $envSettings = $this->storageHandler->loadSettings($envPath);
-            $settings->mergeRecursiveDistinct($envSettings);
+        $settings = new Dot($this->storageHandler->loadSettings($this->settingsFile));
+        $envType = $environment->getType()->toString();
+        foreach ($this->supportedFileTypes as $fileType) {
+            $envPath = sprintf('%s/%s.%s', $this->envDir, $envType, $fileType);
+            if (is_file($envPath)) {
+                $envSettings = $this->storageHandler->loadSettings($envPath);
+                $settings->mergeRecursiveDistinct($envSettings);
+            }
         }
+
         return new DotAccessSettings($settings->all());
     }
 }