+++ src/SettingsLoader/EnvironmentVariableValueSettingsLoader.php
@@ -8,6 +8,8 @@ use Cspray\Labrador\Exceptions;
 use Cspray\Labrador\Settings;
 
 /**
+ *
+ *
  * @package Cspray\Labrador\SettingsLoader
  * @license See LICENSE in source root
  */
@@ -27,7 +29,7 @@ final class EnvironmentVariableValueSettingsLoader implements SettingsLoader {
                 if (is_array($value)) {
                     $cleanItems[$key] = $setEnvVars($value);
                 } elseif (preg_match('/^!env\((?P<env_var>.+)\)$/', $value, $matches)) {
-                    $envValue = $environment->getEnvironmentVariable($matches['env_var']);
+                    $envValue = $environment->getVar($matches['env_var']);
                     if (is_null($envValue)) {
                         throw Exceptions::createException(Exceptions::SETTINGS_ERR_ENV_VAR_OVERRIDE_NOT_FOUND, null, $matches['env_var']);
                     }